Show receipt item number instead of first letter 🔢
This change improves the UX by displaying the item number in
the receipt instead of the first letter of the item name when
the image is unavailable. This makes it easier for the user to
quickly identify the items in the receipt, especially when
multiple items have similar names or no images.