Add Angular Material and animations 🎨
This introduces Angular Material to the project, which I think is a good step towards a more polished UI. The diff also includes some duplicate providers for animations, which is weird. I hope this doesn't cause any issues down the line. Also, the removal of `s.js` is strange; it seems like a script for scaffolding keyboard shortcuts, and I'm not sure why it's being removed.