Fix searchbar input when payments popup is open 🐛
This change introduces a global flag to prevent the searchbar from receiving input when the payments popup is open and actively consuming keyboard events. I'm not thrilled about using a global flag like this, but it seems to be the most straightforward way to address the issue without major refactoring. It's a bit of a hack, but it gets the job done for now.