Add print receipt endpoint for sales 🖨️
This change introduces a new endpoint for printing sales receipts, which seems like a useful feature for businesses that need to provide physical receipts to their customers. The implementation appears straightforward, utilizing puppeteer to render the receipt as an image and then sending it to the printer. However, I'm slightly concerned about the dependency on puppeteer, as it can be resource-intensive and may add complexity to the deployment process.